Mitigate file-hosing CORS issue
CORS does not allow navbar and main document to communicate in a simple way when hosted from files.
Turns out that this can be solved by communicating the active page to the navbar via a document fragment identifer.
This approach should also be extendable to communicating check-states inside navbar etc.